Flush queue on SIGTERM
Introduces 'gracefulfade' status, distinct from 'fade' in that it
will flush the pending syncs queue. This makes for a shutdown
without dataloss -- no new events are accepted, but those already
in the queue are serviced.

We apply gracefulfade when we receive SIGTERM. SIGINT still uses
'fade' and will only wait for already running processes to exit.

Notes:

 - lsyncdStatus becomes a global and is initialized earlier

 - part of runner.cycle() gets split into runner._cycle_syncs()

 - user alarms are not invoked during gracefulfade (should they?)

 - Sync get a new public method countDelays()