Say default for delay; better explainations of using rsync parameter options.
I've often hunted for clear, explicit statements about the default values for the "delay" parameter and I struggled early on to understand the relationship between the rsync parameter options and the sync parameters. So I'm trying to make that stuff more clear. If my changes meet with Axel's approval, perhaps I'll work on others.